# openapi.yaml
openapi: 3.0.3
info:
  title: TSTS - Technical Support Ticketing System API
  description: |
    Comprehensive API documentation for the Technical Support Ticketing System (TSTS).
    
    ## Overview
    TSTS is a full-featured ticketing system designed for university IT support operations.
    It includes ticket management, chat, knowledge base, AI assistance, and comprehensive reporting.
    
    ## Authentication
    The API supports multiple authentication methods:
    - **Bearer Token**: JWT-based authentication for web and mobile clients
    - **API Keys**: For third-party integrations with scope-based access control
    - **WebAuthn**: Passwordless authentication using trusted devices
    
    ## Base URLs
    - Production: `https://api.tsts.local`
    - Development: `http://localhost:3000`
    
    ## Rate Limiting
    - 1000 requests per 15 minutes per IP address
    - Rate limit headers included in responses
    
    ## Pagination
    Most list endpoints support pagination with `page` and `limit` query parameters:
    - Default page: 1
    - Default limit: 20
    - Maximum limit: 200
    
    ## Internationalization
    The API supports English and Arabic. Many fields use JSONB format:
    ```json
    {
      "name": {
        "en": "Computer Science",
        "ar": "علوم الحاسب"
      }
    }
    ```
    
    ## Error Responses
    All errors follow a consistent format:
    ```json
    {
      "message": "Error description",
      "statusCode": 400,
...
    }
    ```
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
    
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for third-party integrations
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          example: "Validation failed"
        statusCode:
          type: integer
          example: 400
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    LocalizedString:
      type: object
      properties:
        en:
          type: string
          example: "Computer Science"
        ar:
          type: string
          example: "علوم الحاسب"
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          $ref: '#/components/schemas/LocalizedString'
        lastName:
          $ref: '#/components/schemas/LocalizedString'
        fullName:
          $ref: '#/components/schemas/LocalizedString'
        user_type:
          type: string
          enum: [Requester, Admin, Technician, SuperAdmin]
        status:
          type: string
          enum: [Active, Inactive]
        image:
          type: string
          format: uri
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Ticket:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ticket_number:
          type: integer
          example: 42
        title:
          type: string
          example: "Cannot access email"
        description:
          type: string
        status:
          type: string
          enum: [draft, open, in_progress, pending, closed, resolved, re_open, out_of_service]
        priority:
          type: string
          enum: [important/urgent, important, urgent, NA]
        requester:
          $ref: '#/components/schemas/User'
        assigneeList:
          type: array
          items:
            $ref: '#/components/schemas/User'
        specialization:
          $ref: '#/components/schemas/Specialization'
        problem:
          $ref: '#/components/schemas/Problem'
        createdAt:
          type: string
          format: date-time
        modifiedAt:
          type: string
          format: date-time
    Specialization:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          $ref: '#/components/schemas/LocalizedString'
        description:
          $ref: '#/components/schemas/LocalizedString'
        review_required:
          type: boolean
    Problem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          $ref: '#/components/schemas/LocalizedString'
        specialization:
          $ref: '#/components/schemas/Specialization'
...
  responses:
...
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: "Resource not found"
            statusCode: 404
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/v1/tickets:
    get:
      summary: List tickets
      tags: [Tickets]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, open, in_progress, pending, closed, resolved, re_open, out_of_service]
        - name: priority
          in: query
          schema:
            type: string
            enum: [important/urgent, important, urgent, NA]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of tickets
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
                  page:
                    type: integer
                  limit:
                    type: integer
                  total:
                    type: integer
                  totalPages:
                    type: integer

    post:
      summary: Create ticket
      tags: [Tickets]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [title, description, priority]
              properties:
                title:
                  type: string
                description:
                  type: string
                priority:
                  type: string
                  enum: [important/urgent, important, urgent, NA]
                departmentId:
                  type: string
                  format: uuid
                specializationId:
                  type: string
                  format: uuid
                media:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '201':
          description: Ticket created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
  /api/v1/tickets/{id}:
    get:
      summary: Get ticket by ID
      tags: [Tickets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ticket'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update ticket (Requester)
      tags: [Tickets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Ticket updated

    delete:
      summary: Delete ticket
      tags: [Tickets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ticket deleted
  /api/v1/tickets/{id}/change-status:
    patch:
      summary: Change ticket status
      tags: [Tickets]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [open, in_progress, pending, closed, resolved, re_open]
                comment:
                  type: string
      responses:
        '200':
          description: Status changed

  # AI Assistant Endpoints
...